+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - refactor: ws client actor logic 
 
+### Fixed
+- don't hang on worth devtools config response 
 
 ## [0.1.2] - 2020-08-16